refactor(viewer): extract OrbitSceneContents shared by Scene and OrbitViewer
The in-Canvas scene graph (camera rig, controls, lights, central body, trails,
satellites) plus its frame/lighting wiring lived only inside the app's Scene,
while the library's OrbitViewer reimplemented a thinner copy. Pull that graph
into a neutral OrbitSceneContents component that both mount through their own
Canvas, so frame handling, lighting and LVLH camera tracking live in one place
instead of being maintained twice.

- OrbitSceneContents.tsx: the scene graph (no Canvas, no global up side effect)
- Scene.tsx: thin Canvas wrapper around it (keeps the app's DEFAULT_UP)
- OrbitViewer: builds the per-satellite maps + internal frame and renders the
  same contents — gaining LVLH camera tracking and the centered-satellite
  rendering (a satellite-centred view now shows the satellite instead of nothing)
- drop the unwired sunDirection override prop (the Sun is computed in the scene)

App behaviour is unchanged: the LVLH-orientation E2E and unit tests stay green.
